+++ R/ee_as_sf.R
@@ -80,7 +80,7 @@
 #'
 #' For getting more information about exporting data from Earth Engine, take
 #' a look at the
-#' \href{https://developers.google.com/earth-engine/exporting}{Google
+#' \href{https://developers.google.com/earth-engine/exporting/}{Google
 #' Earth Engine Guide - Export data}.
 #' @return An sf object.
 #' @family vector download functions